Rename Spectral files to SpectralInstrument
