+++ include/SpectralInstrumentCompatibility.h
