+++ src/SpectralInstrumentCameraBin.hpp
