+++ src/SpectralInstrumentCameraDetInfo.hpp
