+++ src/SpectralInstrumentCameraInterface.hpp
