+++ src/SpectralInstrumentCameraRoi.hpp
