+++ src/SpectralInstrumentCameraSync.hpp
