+++ src/SpectralInstrumentDetInfoCtrlObj.cpp
@@ -19,7 +19,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program; if not, see <http://www.gnu.org/licenses/>.
 //###########################################################################
-#include "SpectralDetInfoCtrlObj.h"
+#include "SpectralInstrumentDetInfoCtrlObj.h"
 
 using namespace lima;
 using namespace lima::Spectral;
