+++ src/SpectralInstrumentInterface.cpp
@@ -19,7 +19,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program; if not, see <http://www.gnu.org/licenses/>.
 //###########################################################################
-#include "SpectralInterface.h"
+#include "SpectralInstrumentInterface.h"
 #include <algorithm>
 
 using namespace lima;
