+++ src/SpectralInstrumentSyncCtrlObj.cpp
@@ -20,7 +20,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program; if not, see <http://www.gnu.org/licenses/>.
 //###########################################################################
-#include "SpectralSyncCtrlObj.h"
+#include "SpectralInstrumentSyncCtrlObj.h"
 
 using namespace lima;
 using namespace lima::Spectral;
